+++ transactions/src/main/java/nl/tudelft/sem/transactions/controllers/TransactionController.java
@@ -10,6 +10,7 @@ import nl.tudelft.sem.transactions.repositories.ProductRepository;
 import nl.tudelft.sem.transactions.repositories.TransactionsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -51,39 +52,43 @@ public class TransactionController {
      */
     @PostMapping("/addNewTransaction")
     public @ResponseBody
-    boolean addNewTransaction(@RequestBody Transactions transaction) {
+    ResponseEntity<String> addNewTransaction(@RequestBody Transactions transaction) {
         
-        Product product = productRepository.findByProductId(transaction.getProductId());
-        if (product == null) {
-            return false;
+        Optional<Product> optionalProduct = productRepository.findByProductId(
+                transaction.getProductId());
+
+        if (optionalProduct.isPresent()) {
+            Product product = optionalProduct.get();
+            int portionsLeft = product.getPortionsLeft()
+                    - transaction.getPortionsConsumed();
+
+            if (transaction.getProductFk().getExpired() == 1 || portionsLeft < 0) {
+                return ResponseEntity.badRequest().body(
+                        "Product is expired or there is no portions left");
+            }
+
+            float credits = product.getPrice()
+                    / product.getTotalPortions();
+
+            credits = credits * transaction.getPortionsConsumed();
+            credits = Math.round(credits * 100) / 100;
+
+
+            try {
+                transactionsRepository.save(transaction);
+                productRepository.updateExistingProduct(product.getProductName(),
+                        product.getUsername(), product.getPrice(), product.getTotalPortions(),
+                        portionsLeft, 0, product.getProductId());
+
+                MicroserviceCommunicator.sendRequestForChangingCredits(transaction.getUsername(),
+                        credits, false);
+                return ResponseEntity.ok().body("Transaction was successfully added");
+            } catch (DataIntegrityViolationException e) {
+                return ResponseEntity.badRequest().body("Adding the transaction failed");
+            }
+
         }
-        
-        int portionsLeft = product.getPortionsLeft()
-                                   - transaction.getPortionsConsumed();
-        
-        if (transaction.getProductFk().getExpired() == 1 || portionsLeft < 0) {
-            return false;
-        }
-        
-        float credits = product.getPrice()
-                                / product.getTotalPortions();
-        
-        credits = credits * transaction.getPortionsConsumed();
-        credits = Math.round(credits * 100) / 100;
-        
-    
-        try {
-            transactionsRepository.save(transaction);
-            productRepository.updateExistingProduct(product.getProductName(),
-                    product.getUsername(), product.getPrice(), product.getTotalPortions(),
-                    portionsLeft, 0, product.getProductId());
-            
-            MicroserviceCommunicator.sendRequestForChangingCredits(transaction.getUsername(),
-                    credits, false);
-            return true;
-        } catch (DataIntegrityViolationException e) {
-            return false;
-        }
+        return ResponseEntity.notFound().build();
     }
 
     /**New transaction which credits will be split among the users that are eating together.
@@ -94,46 +99,49 @@ public class TransactionController {
      */
     @PostMapping("/transactionSplittingCredits")
     public @ResponseBody
-    boolean addNewTransactionSplittingCredits(@RequestBody TransactionsSplitCredits
+    ResponseEntity<String> addNewTransactionSplittingCredits(@RequestBody TransactionsSplitCredits
                                                       transactionsSplitCredits) {
         
         
         Transactions transaction = transactionsSplitCredits.getTransactionsSplit();
         
-        Product product = productRepository.findByProductId(transaction.getProductId());
+        Optional<Product> optionalProduct = productRepository.findByProductId(
+                transaction.getProductId());
         
-        if (product == null) {
-            return false;
+        if (optionalProduct.isPresent()) {
+            Product product = optionalProduct.get();
+            int portionsLeft = product.getPortionsLeft()
+                    - transaction.getPortionsConsumed();
+
+            if (product.getExpired() == 1 || portionsLeft < 0) {
+                return ResponseEntity.badRequest().body(
+                        "Product is expired or there is no portions left");
+            }
+
+            List<String> usernames = transactionsSplitCredits.getUsernames();
+            float credits = product.getPrice()
+                    / product.getTotalPortions();
+
+            credits = credits * transaction.getPortionsConsumed();
+            float splitCredits = credits / usernames.size();
+
+            splitCredits = Math.round(splitCredits * 100) / 100;
+
+            try {
+                productRepository.updateExistingProduct(product.getProductName(),
+                        product.getUsername(), product.getPrice(),
+                        product.getTotalPortions(), portionsLeft, 0, product.getProductId());
+
+                transactionsRepository.save(transaction);
+                MicroserviceCommunicator.sendRequestForSplittingCredits(usernames, splitCredits);
+
+                return ResponseEntity.ok().body("Transaction was successfully added");
+            } catch (DataIntegrityViolationException e) {
+                return ResponseEntity.badRequest().body("Adding the transaction failed");
+            }
         }
         
-        int portionsLeft = product.getPortionsLeft()
-                                   - transaction.getPortionsConsumed();
-    
-        if (product.getExpired() == 1 || portionsLeft < 0) {
-            return false;
-        }
-    
-        List<String> usernames = transactionsSplitCredits.getUsernames();
-        float credits = product.getPrice()
-                                / product.getTotalPortions();
-        
-        credits = credits * transaction.getPortionsConsumed();
-        float splitCredits = credits / usernames.size();
-        
-        splitCredits = Math.round(splitCredits * 100) / 100;
-        
-        try {
-            productRepository.updateExistingProduct(product.getProductName(),
-                    product.getUsername(), product.getPrice(),
-                    product.getTotalPortions(), portionsLeft, 0, product.getProductId());
-            
-            transactionsRepository.save(transaction);
-            MicroserviceCommunicator.sendRequestForSplittingCredits(usernames, splitCredits);
-            
-            return true;
-        } catch (DataIntegrityViolationException e) {
-            return false;
-        }
+        return ResponseEntity.notFound().build();
     }
 
     /**
@@ -205,16 +213,19 @@ public class TransactionController {
      */
     @DeleteMapping("/deleteTransaction/{transactionId}")
     @ResponseBody
-    public void deleteTransaction(@PathVariable(value = "transactionId") long transactionId) {
-        try {
-            Optional<Transactions> t = transactionsRepository.findById(transactionId);
-            Transactions transaction = t.get();
-            transaction.getProductFk().removeTransaction(transaction);
-            transaction.setProductFk(null);
-            transactionsRepository.delete(transaction);
-            System.out.println("The transaction was deleted.");
-        } catch (Exception e) {
-            throw new IllegalArgumentException("the deletion has failed");
+    public ResponseEntity<String> deleteTransaction(@PathVariable long transactionId) {
+        Optional<Transactions> optTransactions = transactionsRepository.findById(transactionId);
+        if (optTransactions.isPresent()) {
+            try {
+                Transactions transaction = optTransactions.get();
+                transaction.getProductFk().removeTransaction(transaction);
+                transaction.setProductFk(null);
+                transactionsRepository.delete(transaction);
+                return ResponseEntity.ok().body("Transaction successfully deleted.");
+            } catch (Exception e) {
+                throw new IllegalArgumentException("the deletion has failed");
+            }
         }
+        return ResponseEntity.notFound().build();
     }
 }
